+++ src/cache/contexts.ts
@@ -1,6 +1,6 @@
 import { MessageContextMenuCommandInteraction, UserContextMenuCommandInteraction } from 'discord.js'
 import { CacheHandler, registerCache } from '../utils.js'
-import { InteractionHandler } from '../types.js'
+import type { InteractionHandler } from '../types.js'
 import { join } from 'node:path'
 
 export default await registerCache(