+++ src/cache/commands.ts
@@ -1,6 +1,5 @@
-import { CacheHandler } from '../utils.js'
+import { CacheHandler, InteractionHandler } from '../utils.js'
 import { ChatInputCommandInteraction } from 'discord.js'
-import type { InteractionHandler } from '../types.js'
 import { join } from 'node:path'
 
 export default await new CacheHandler<InteractionHandler<ChatInputCommandInteraction>>().register(