+++ src/cache/contexts.ts
@@ -1,6 +1,5 @@
 import { MessageContextMenuCommandInteraction, UserContextMenuCommandInteraction } from 'discord.js'
-import { CacheHandler } from '../utils.js'
-import type { InteractionHandler } from '../types.js'
+import { CacheHandler, InteractionHandler } from '../utils.js'
 import { join } from 'node:path'
 
 export default await new CacheHandler<