+++ src/index.ts
@@ -5,9 +5,8 @@ import {
     Events,
     IntentsBitField,
 } from 'discord.js'
+import { formatName, registerEvents, InteractionFile } from './utils.js'
 import autocompleteCache from './cache/autocompletes.js'
-import { formatName, registerEvents } from './utils.js'
-import type { InteractionFile } from './types.js'
 import commandsCache from './cache/commands.js'
 import contextCache from './cache/contexts.js'
 import buttonsCache from './cache/buttons.js'